Enable IMPLICIT_CONVERSION_LOSS and IMPLICIT_INTEGER_CAST_LOSS for ISSigTool + fix all warnings that causes.
The settings are still in the .dproj because they're actually disabled by default in Delphi.

The DWORD(E_OUTOFMEMORY) code seems odd to me, but does work.

Includes a cleanup of no longer needed NTSTATUS type definition.

+++ Projects/Src/Compression.SevenZipDecoder.pas
@@ -317,7 +317,7 @@ procedure Extract7ZipArchiveRedir(const DisableFsRedir: Boolean;
       SZ_ERROR_DATA, SZ_ERROR_CRC, SZ_ERROR_ARCHIVE:
         SevenZipError(SetupMessages[msgArchiveIsCorrupted]);
       SZ_ERROR_MEM:
-        SevenZipError(Win32ErrorString(E_OUTOFMEMORY));
+        SevenZipError(Win32ErrorString(DWORD(E_OUTOFMEMORY)));
     else
       SevenZipError(Res.ToString);
     end;
